Tidying up a little bit
