+++ src/control_unit/control_unit/time.h
@@ -1,0 +1,16 @@
+
+struct Time{
+  int year;
+  int month;
+  int day;
+  int hours;
+  int minutes;
+  int seconds;
+};
+
+// Retrievs real time from Ethernet
+// @return time in seconds
+Time getRealTime();
+int getYear();
+int getMonth();
+int getDay();
